2019-03-02

+++ src/Inventory/InputService.js
@@ -1,4 +1,4 @@
-import {inject, injectable} from "inversify";
+import { inject, injectable } from "inversify";
 
 @injectable()
 export class InputService {
@@ -15,10 +15,22 @@ export class InputService {
     @inject("InventoryControlTurnoverService")
     /**@type{InventoryControlTurnoverService}*/ inventoryControlTurnoverService = undefined;
 
+    @inject("InventoryPricingRepository")
+    /**@type {InventoryPricingRepository}*/ inventoryPricingRepository = undefined;
+
+    @inject("JournalGenerationTemplateService")
+    /**@type{JournalGenerationTemplateService}*/ journalGenerationTemplateService = undefined;
+
+    @inject("InventoryIOTypeRepository")
+    /**@type{InventoryIOTypeRepository}*/ inventoryIOTypeRepository = undefined;
+
     _mapToEntity(cmd) {
         return {
             id: cmd.id,
+            time: cmd.id ? undefined : ( cmd.time || new Date ),
             stockId: cmd.stockId,
+            sourceStockId: cmd.sourceStockId,
+            delivererId: cmd.delivererId,
             invoiceId: cmd.invoiceId,
             date: cmd.date || Utility.PersianDate.current(),
             description: cmd.description,
@@ -26,13 +38,16 @@ export class InputService {
             quantityStatus: 'draft',
             ioType: cmd.ioType,
             fiscalPeriodId: this.state.fiscalPeriodId,
-            inventoryLines: (cmd.inventoryLines || cmd.lines).asEnumerable()
-                .select(line => ({
+            inventoryLines: ( cmd.inventoryLines || cmd.lines ).asEnumerable()
+                .select(line => ( {
                     id: cmd.id,
                     productId: line.productId,
+                    baseInventoryId: line.baseInventoryId,
                     quantity: line.quantity,
-                    unitPrice: 0
-                }))
+                    unitPrice: line.unitPrice || 0,
+                    vat: line.vat,
+                    tax: line.tax
+                } ))
                 .toArray()
         }
     }
@@ -57,9 +72,25 @@ export class InputService {
             if (Utility.String.isNullOrEmpty(line.productId))
                 errors.push(`کالا وجود ندارد - ردیف {0}`.format(row));
 
-            if (!(line.quantity && line.quantity > 0))
+            if (!( line.quantity && line.quantity > 0 ))
                 errors.push(`مقدار وجود ندارد - ردیف {0}`.format(row));
         });
+
+        const lastPrice = this.inventoryPricingRepository.findLast();
+
+        if (!cmd.id) {
+            if (lastPrice && cmd.date < lastPrice.toDate)
+                throw new ValidationException([ 'صدور رسید در تاریخ قبل از قیمت گذاری امکان پذیر نیست' ]);
+        }
+        else {
+            const input = this.inventoryRepository.findById(cmd.id);
+
+            if (lastPrice && input.date < lastPrice.toDate)
+                throw new ValidationException([ 'تغییر رسید در تاریخ های قبل از قیمت گذاری امکان پذیر نیست' ]);
+
+            if (input.journalId)
+                throw new ValidationException([ 'برای رسید جاری سند صادر شده امکان ویرایش وجود ندارد' ]);
+        }
 
         return errors;
     }
@@ -92,7 +123,7 @@ export class InputService {
         if (!input)
             throw new NotFoundException();
 
-        this.inventoryRepository.update(id, {quantityStatus: 'confirmed'});
+        this.inventoryRepository.update(id, { quantityStatus: 'confirmed' });
 
         this.eventBus.send("InventoryInputCreated", id);
     }
@@ -104,7 +135,7 @@ export class InputService {
         if (!input)
             throw new NotFoundException();
 
-        this.inventoryRepository.update(id, {quantityStatus: 'fixed'});
+        this.inventoryRepository.update(id, { quantityStatus: 'fixed' });
 
         this.eventBus.send("InventoryInputFixed", id);
     }
@@ -120,15 +151,15 @@ export class InputService {
             throw new ValidationException(errors);
 
         if (input.quantityStatus === 'fixed')
-            throw new ValidationException(['رسید ثبت قطعی شده ، امکان تغییر وجود ندارد']);
+            throw new ValidationException([ 'رسید ثبت قطعی شده ، امکان تغییر وجود ندارد' ]);
 
         let removedLines = input.inventoryLines.asEnumerable()
             .where(inputLine => !cmd.inventoryLines.asEnumerable().any(line => line.id === inputLine.id))
-            .select(item => ({
+            .select(item => ( {
                 id: item.id,
                 productId: item.productId,
                 quantity: 0
-            }))
+            } ))
             .toArray();
 
         errors = this.inventoryControlTurnoverService.validateTurnover({
@@ -141,11 +172,11 @@ export class InputService {
                 /* removed lines */
                 .concat(removedLines)
 
-                .select(item => ({
+                .select(item => ( {
                     id: item.id,
                     productId: item.productId,
                     quantity: item.quantity
-                }))
+                } ))
                 .toArray()
         });
 
@@ -155,7 +186,7 @@ export class InputService {
         let entity = {
             number: cmd.ioType === input.ioType && cmd.stockId === input.stockId
                 ? input.number
-                : (this.inventoryRepository.inputMaxNumber(this.state.fiscalPeriodId, cmd.stockId, cmd.ioType).max || 0) + 1,
+                : ( this.inventoryRepository.inputMaxNumber(this.state.fiscalPeriodId, cmd.stockId, cmd.ioType).max || 0 ) + 1,
             date: cmd.date || Utility.PersianDate.current(),
             stockId: cmd.stockId,
             inventoryType: 'input',
@@ -164,11 +195,11 @@ export class InputService {
         };
 
         entity.inventoryLines = cmd.inventoryLines.asEnumerable()
-            .select(line => ({
+            .select(line => ( {
                 productId: line.productId,
                 quantity: line.quantity,
                 unitPrice: line.unitPrice
-            })).toArray();
+            } )).toArray();
 
         this.inventoryRepository.updateBatch(id, entity);
 
@@ -188,11 +219,11 @@ export class InputService {
             errors = this.inventoryControlTurnoverService.validateTurnover({
                 stockId: input.stockId,
                 inventoryLines: input.inventoryLines.asEnumerable()
-                    .select(item => ({
+                    .select(item => ( {
                         id: item.id,
                         productId: item.productId,
                         quantity: 0
-                    }))
+                    } ))
                     .toArray()
             });
 
@@ -200,7 +231,13 @@ export class InputService {
             throw new ValidationException(errors);
 
         if (input.quantityStatus === 'fixed')
-            throw new ValidationException(['حواله ثبت قطعی شده ، امکان حذف وجود ندارد']);
+            throw new ValidationException([ 'رسید ثبت قطعی شده ، امکان حذف وجود ندارد' ]);
+
+        if(input.journalId)
+            throw new ValidationException(['برای رسید جاری سند حسابداری صادر شده ، ابتدا سند حسابداری را حذف کنید ']);
+
+        if (this.inventoryPricingRepository.isInventoryExist(id))
+            throw new ValidationException([ 'رسید جاری در قیمت گذاری وجود دارد ، امکان حذف وجود ندارد' ]);
 
         this.inventoryRepository.remove(id);
 
@@ -211,8 +248,37 @@ export class InputService {
 
         const input = this.inventoryRepository.findById(id);
 
-        this.inventoryRepository.update(id, {invoiceId, quantityStatus: 'confirmed'});
+        this.inventoryRepository.update(id, { invoiceId, quantityStatus: 'confirmed' });
 
         this.eventBus.send("InventoryInputChanged", input, id);
+    }
+
+    generateJournal(id) {
+        let input = this.inventoryRepository.findById(id);
+
+        if (!input)
+            throw new NotFoundException();
+
+        if (input.journalId)
+            throw new ValidationException([ 'سند حسابداری قبلا صادر شده' ]);
+
+        const hasZeroPrice = input.inventoryLines.asEnumerable().any(item => item.unitPrice === 0);
+
+        if (hasZeroPrice)
+            throw new ValidationException([ 'رسید جاری قیمت گذاری نشده ، امکان صدور سند حسابداری وجود ندارد' ]);
+
+        const ioType = input.ioType
+            ? this.inventoryIOTypeRepository.findById(input.ioType)
+            : null;
+
+        if (!ioType)
+            throw new ValidationException([ 'رسید دارای نوع نمی باشد ، امکان صدور سند وجود ندارد' ]);
+
+        if (!ioType.journalGenerationTemplateId)
+            throw new ValidationException([ 'برای نوع رسید جاری الگوی صدور سند حسابداری وجود ندارد ، امکان صدور سند وجود ندارد' ]);
+
+        const journalId = this.journalGenerationTemplateService.generate(ioType.journalGenerationTemplateId, id, 'Inventory');
+
+        this.inventoryRepository.update(id, { journalId });
     }
 }
